Fix GT response so it matches spec
The Moonlite protocol spec indicates that the firmware should respond to the get
temperature command "GT" with an ASCII-encoded 16-bit signed integer. Without
this change the INDI moonlight driver is emitting error messages about the
gettemperature command timing out. This is because the INDI driver is waiting to
read 5 characters and it's only getting 3.